UI updating.
Now the UI is fully rebuilt by PySide2, and the front UI is separated from the backward programs.

Accordingly, several changes has been made to different channels such that they are consistent with the concept design.

Sample Channels have been moved to a different directory, which is "parallel" with other packages in Pynars.

Next, I will make the UI looks better and then work for an "ultimate" working environment.